some refact
